增加测试霍尔的方法

+++ LC_Fixture/APP/hal.h
@@ -55,7 +55,7 @@ struct _cmd_list{
 };
 struct _obj{
 	u8 enable;
-	u8 state;
+	int state;
 	u32 tmp;
 };
 enum{
